Changed behavior of normalize to cause it to always override port and user with values from ssh config.
This will require those who wish to override ssh config values to do
so in the host_string itself.  Before this, if the port or user were
set for one host then execute(..., hosts=[x,y,z]) was used, the port
and user would only be taken from the ssh config if these weren't set
for the current host in the master task.

The only options were to save and clear these values before calling
execute or to specify the port or user in the host_string along with
the hostname from ssh config which removed a lot of the purpose of
use_ssh_config.